+++ projectroles/forms.py
@@ -3,10 +3,11 @@ import json
 from django import forms
 from django.conf import settings
 from django.contrib import auth
+from django.urls import reverse
 from django.utils import timezone
 
 from pagedown.widgets import PagedownWidget
-from dal import autocomplete
+from dal import autocomplete, forward as dal_forward
 
 from .models import (
     Project,
@@ -52,14 +53,14 @@ DELEGATE_LIMIT = getattr(settings, 'PROJECTROLES_DELEGATE_LIMIT', 1)
 User = auth.get_user_model()
 
 
-# General widgets --------------------------------------------------------------
+# User autocompletion ----------------------------------------------------------
 
 
-class UserAutocompleteWidget(autocomplete.ModelSelect2):
-    """Custom Select widget for user field autocompletion that uses the UUID
-    instead of the pk."""
+class SODARUserAutocompleteWidget(autocomplete.ModelSelect2):
+    """Custom Select widget for user field autocompletion which uses the SODAR
+    User class sodar_uuid instead of pk"""
 
-    # override function to use sodar_uuid instead of pk
+    # Override function to use sodar_uuid instead of pk
     def filter_choices_to_render(self, selected_choices):
         """Filter out un-selected choices if choices is a QuerySet."""
         self.choices.queryset = self.choices.queryset.filter(
@@ -67,10 +68,102 @@ class UserAutocompleteWidget(autocomplete.ModelSelect2):
         )
 
 
-class RedirectWidget(UserAutocompleteWidget):
-    """Custom Select widget for QuerySet choices and Select2."""
+class SODARUserRedirectWidget(SODARUserAutocompleteWidget):
+    """Version of SODARUserAutocompleteWidget to allow redirect on selection"""
 
     autocomplete_function = 'autocomplete_redirect'
+
+
+# TODO: Refactor into widget?
+def get_user_widget(
+    scope='all',
+    project=None,
+    exclude=None,
+    forward=None,
+    url=None,
+    widget_class=None,
+):
+    """
+    Get an user autocomplete widget for your form.
+
+    :param scope: Scope of users to include: "all"/"project"/"project_exclude"
+    :param project: Project object or project UUID string (optional)
+    :param exclude: List of User objects or User UUIDs to exclude
+    :param forward: Parameters to forward to autocomplete view (optional)
+    :param url: Autocomplete ajax class override (optional)
+    :param widget_class: Widget class override (optional)
+    :return: SODARUserAutocompleteWidget or an overridden widget class
+    """
+    url = url or 'projectroles:autocomplete_user'
+    wg = {'url': url, 'forward': [dal_forward.Const(scope, 'scope')]}
+
+    if project:
+        p_uuid = (
+            str(project.sodar_uuid)
+            if isinstance(project, Project)
+            else str(project)
+        )
+        wg['forward'].append(dal_forward.Const(p_uuid, 'project'))
+
+    if forward and isinstance(forward, list):
+        wg['forward'] += forward
+
+    if exclude:
+        wg['forward'].append(
+            dal_forward.Const(
+                [
+                    str(u.sodar_uuid) if isinstance(u, User) else u
+                    for u in exclude
+                ],
+                'exclude',
+            )
+        )
+
+    if widget_class:
+        return widget_class(**wg)
+
+    return SODARUserAutocompleteWidget(**wg)
+
+
+class SODARUserChoiceField(forms.ModelChoiceField):
+    """User choice field to be used with SODAR User objects and autocomplete"""
+
+    def __init__(
+        self,
+        scope='all',
+        project=None,
+        exclude=None,
+        forward=None,
+        url=None,
+        widget_class=None,
+        *args,
+        **kwargs
+    ):
+        """
+        Override of ModelChoiceField initialization.
+
+        Most arguments given to ModelChoiceField can be set, with the exception
+        of queryset, to_field_name, limit_choices_to and widget.
+
+        :param scope: Scope of users to include:
+                      "all"/"project"/"project_exclude"
+        :param project: Project object or project UUID string (optional)
+        :param exclude: List of User objects or User UUIDs to exclude
+        :param forward: Parameters to forward to autocomplete view (optional)
+        :param url: Autocomplete ajax class override (optional)
+        :param widget_class: Widget class override (optional)
+        """
+        widget = get_user_widget(
+            scope, project, exclude, forward, url, widget_class
+        )
+        super().__init__(
+            User.objects.all(),
+            to_field_name='sodar_uuid',
+            limit_choices_to=None,
+            widget=widget,
+            *args,
+            **kwargs
+        )
 
 
 # Project form -----------------------------------------------------------------
@@ -79,25 +172,8 @@ class RedirectWidget(UserAutocompleteWidget):
 class ProjectForm(forms.ModelForm):
     """Form for Project creation/updating"""
 
-    owner = forms.ModelChoiceField(
-        User.objects.all(),
-        required=True,
-        to_field_name='sodar_uuid',
-        label='Owner',
-        help_text='Owner',
-        widget=UserAutocompleteWidget(
-            url='projectroles:autocomplete_user_exclude', forward=['project']
-        ),
-    )
-
-    # Hidden project field for user autocomplete
-    project = forms.ModelChoiceField(
-        Project.objects.all(),
-        required=False,
-        to_field_name='sodar_uuid',
-        initial=None,
-        widget=forms.HiddenInput(),
-    )
+    # Set up owner field
+    owner = SODARUserChoiceField(label='Owner', help_text='Owner')
 
     class Meta:
         model = Project
@@ -366,12 +442,6 @@ class RoleAssignmentForm(forms.ModelForm):
     class Meta:
         model = RoleAssignment
         fields = ['project', 'user', 'role']
-        widgets = {
-            'user': RedirectWidget(
-                url='projectroles:autocomplete_user_redirect',
-                forward=['project', 'role'],
-            )
-        }
 
     def __init__(self, project=None, current_user=None, *args, **kwargs):
         """Override for form initialization"""
@@ -394,9 +464,17 @@ class RoleAssignmentForm(forms.ModelForm):
         # Form modifications
         ####################
 
-        # Modify ModelChoiceFields to use sodar_uuid
+        # Modify project field to use sodar_uuid
         self.fields['project'].to_field_name = 'sodar_uuid'
-        self.fields['user'].to_field_name = 'sodar_uuid'
+
+        # Set up user field
+        self.fields['user'] = SODARUserChoiceField(
+            scope='project_exclude',
+            project=self.project,
+            forward=['role'],
+            url=reverse('projectroles:autocomplete_user_redirect'),
+            widget_class=SODARUserRedirectWidget,
+        )
 
         # Limit role choices
         self.fields['role'].choices = get_role_choices(
@@ -493,18 +571,16 @@ class RoleAssignmentOwnerTransferForm(forms.Form):
         self.project = project
         self.current_owner = current_owner
 
-        # ------------------------------
-        self.selectable_roles = get_role_choices(
-            self.project, self.current_user
-        )
-        self.fields['new_owner'] = forms.ModelChoiceField(
+        self.fields['new_owner'] = SODARUserChoiceField(
             label='New owner',
             help_text='Select a member of the project to become owner.',
-            queryset=User.objects.all(),
-            to_field_name='sodar_uuid',
-            widget=UserAutocompleteWidget(
-                url='projectroles:autocomplete_user', forward=['project']
-            ),
+            scope='project',
+            project=self.project,
+            exclude=[self.current_owner],
+        )
+
+        self.selectable_roles = get_role_choices(
+            self.project, self.current_user
         )
         self.fields['ex_owner_role'] = forms.ChoiceField(
             label='New role for {}'.format(self.current_owner.username),